update to master

+++ src/api/controler/RoomController.js
@@ -11,8 +11,10 @@ const { idValidate } = require('../validate/syntaxCheck');
 exports.initRoom = async (data) => {
     const { _id, otherUser } = data;
 
-    idValidate(_id, "init_room_res");
-    idValidate(otherUser, "init_room_res");
+    if (!idValidate(_id, "init_room_res") ||
+        !idValidate(otherUser, "init_room_res")) {
+        throw new Error("Incorrect id form");
+    }
 
     await initRoomService(_id, otherUser);
 };
@@ -20,8 +22,10 @@ exports.initRoom = async (data) => {
 exports.sendMessage = async (data) => {
     const { _id, room, message } = data;
 
-    idValidate(_id, "send_msg_res");
-    idValidate(room, "send_msg_res");
+    if (!idValidate(_id, "send_msg_res") ||
+        !idValidate(room, "send_msg_res")) {
+        throw new Error("Incorrect id form");
+    }
 
     await sendMessageService(_id, room, message);
 };
@@ -29,8 +33,10 @@ exports.sendMessage = async (data) => {
 exports.joinRoom = async (data) => {
     const { _id, room_id } = data;
 
-    idValidate(_id, "init_room_res");
-    idValidate(room_id, "init_room_res");
+    if (!idValidate(_id, "init_room_res") ||
+        !idValidate(room_id, "init_room_res")) {
+        throw new Error("Incorrect id form");
+    }
 
     await joinRoomService(_id, room_id);
 }
@@ -38,8 +44,10 @@ exports.joinRoom = async (data) => {
 exports.readBy = async (data) => {
     const { _id, room_id } = data;
 
-    idValidate(_id, "read_msg_res");
-    idValidate(room_id, "read_msg_res");
+    if (!idValidate(_id, "read_msg_res") ||
+        !idValidate(room_id, "read_msg_res")) {
+        throw new Error("Incorrect id form");
+    }
 
     await readByService(_id, room_id);
 };
@@ -47,9 +55,11 @@ exports.readBy = async (data) => {
 exports.deleteMsg = async (data) => {
     const { _id, room_id, message_id } = data;
 
-    idValidate(_id, "delete_msg_res");
-    idValidate(room_id, "delete_msg_res");
-    idValidate(message_id, "delete_msg_res");
+    if (!idValidate(_id, "delete_msg_res") ||
+        !idValidate(room_id, "delete_msg_res") ||
+        !idValidate(message_id, "delete_msg_res")) {
+        throw new Error("Incorrect id form");
+    }
 
     await deleteMsgService(_id, room_id, message_id);
 };
@@ -57,8 +67,10 @@ exports.deleteMsg = async (data) => {
 exports.deleteChat = async (data) => {
     const { _id, room_id } = data;
 
-    idValidate(_id, "read_msg_res");
-    idValidate(room_id, "read_msg_res");
+    if (!idValidate(_id, "read_msg_res") ||
+        !idValidate(room_id, "read_msg_res")) {
+        throw new Error("Incorrect id form");    
+    }
 
     await deleteChatService(_id, room_id);
 };
